chore: more shared
